BasicDerivatives

Description

Plain functions

acos' :: Double -> Double -> Maybe Double

Math f(x) = acos(x)

acosh' :: Double -> Double -> Maybe Double

Math f(x) = acosh(x)

acot :: Double -> Double -> Maybe Double

Math f(x) = acot(x)

acoth :: Double -> Double -> Maybe Double

Math f(x) = acoth(x)

acsc :: Double -> Double -> Maybe Double

Math f(x) = acsc(x)

acsch :: Double -> Double -> Maybe Double

Math f(x) = acsch(x)

asec :: Double -> Double -> Maybe Double

Math f(x) = asec(x)

asech :: Double -> Double -> Maybe Double

Math f(x) = asech(x)

asin' :: Double -> Double -> Maybe Double

Math f(x) = asin(x)

asinh' :: Double -> Double -> Maybe Double

Math f(x) = asinh(x)

atan' :: Double -> Double -> Maybe Double

Math f(x) = atan(x)

atanh' :: Double -> Double -> Maybe Double

Math f(x) = atanh(x)

aToX :: Double -> Double -> Maybe Double

Math f(a,x) = a^x

cos' :: Double -> Double -> Maybe Double

Math f(x) = cos(x)

cosh' :: Double -> Double -> Maybe Double

Math f(x) = cosh(x)

cot :: Double -> Double -> Maybe Double

Math f(x) = cot(x)

coth :: Double -> Double -> Maybe Double

Math f(x) = coth(x)

csc :: Double -> Double -> Maybe Double

Math f(x) = csc(x)

csch :: Double -> Double -> Maybe Double

Math f(x) = csch(x)

exp' :: Double -> Double -> Maybe Double

Math f(x) = exp(x)

log' :: Double -> Double -> Maybe Double

Math f(x) = loge(x)

loga :: Double -> Double -> Maybe Double

Math f(a,x) = loga(x)

sec :: Double -> Double -> Maybe Double

Math f(x) = sec(x)

sech :: Double -> Double -> Maybe Double

Math f(x) = sech(x)

sin' :: Double -> Double -> Maybe Double

Math f(x) = sin(x)

sinh' :: Double -> Double -> Maybe Double

Math f(x) = sinh(x)

tan' :: Double -> Double -> Maybe Double

Math f(x) = tan(x)

tanh' :: Double -> Double -> Maybe Double

Math f(x) = tanh(x)

xToN :: Double -> Double -> Maybe Double

Math f(n,x) = x^n

xToX :: Double -> Double -> Maybe Double

Math f(x) = x^x

Derivatives of functions

dacos :: Double -> Double -> Maybe Double

Math d(acos(x))/dx

dacosh :: Double -> Double -> Maybe Double

Math d(acosh(x))/dx

dacot :: Double -> Double -> Maybe Double

Math d(acot(x))/dx

dacoth :: Double -> Double -> Maybe Double

Math d(acoth(x))/dx

dacsc :: Double -> Double -> Maybe Double

Math d(acsc(x))/dx

dacsch :: Double -> Double -> Maybe Double

Math d(acsch(x))/dx

dasec :: Double -> Double -> Maybe Double

Math d(asec(x))/dx

dasech :: Double -> Double -> Maybe Double

Math d(asech(x))/dx

dasin :: Double -> Double -> Maybe Double

Math d(asin(x))/dx

dasinh :: Double -> Double -> Maybe Double

Math d(asinh(x))/dx

datan :: Double -> Double -> Maybe Double

Math d(atan(x))/dx

datanh :: Double -> Double -> Maybe Double

Math d(atanh(x))/dx

daToX :: Double -> Double -> Maybe Double

Math d(a^x(a,x))/dx

dc :: Double -> Double -> Maybe Double

Math dc/dx

dcos :: Double -> Double -> Maybe Double

Math d(cos(x))/dx

dcosh :: Double -> Double -> Maybe Double

Math d(cosh(x))/dx

dcot :: Double -> Double -> Maybe Double

Math d(cot(x))/dx

dcoth :: Double -> Double -> Maybe Double

Math d(coth(x))/dx

dloga :: Double -> Double -> Maybe Double

Math d(loga(a,x))/dx

dcsc :: Double -> Double -> Maybe Double

Math d(csc(x))/dx

dcsch :: Double -> Double -> Maybe Double

Math d(csch(x))/dx

dexp :: Double -> Double -> Maybe Double

Math d(exp(x))/dx

dlog :: Double -> Double -> Maybe Double

Math d(loge(x))/dx

dsec :: Double -> Double -> Maybe Double

Math d(sec(x))/dx

dsech :: Double -> Double -> Maybe Double

Math d(sech(x))/dx

dsin :: Double -> Double -> Maybe Double

Math d(sin(x))/dx

dsinh :: Double -> Double -> Maybe Double

Math d(sinh(x))/dx

dtan :: Double -> Double -> Maybe Double

Math d(tan(x))/dx

dtanh :: Double -> Double -> Maybe Double

Math d(tanh(x))/dx

dxToN :: Double -> Double -> Maybe Double

Math d(x^n(n,x))/dx

dxToX :: Double -> Double -> Maybe Double

Math d(x^x(x))/dx

Opening Maybe functions

takeMaybeTD :: Maybe Double -> Double

Function which open Maybe monad, when we time or divide

takeMaybePM :: Maybe Double -> Double

Function which open Maybe monad, when we plus or minus