Main

Description

Main Functions

main :: IO ()

Function which prints welcoming text and calls parseFunctionsNames

parseFunctionsNames :: IO ()

Function which depending on length of argument line calls appropriate function

parseArgumentsName :: [Char] -> IO ()

Function which takes argument and if there's a valid number it calculate the result and print

Other Functions

parseNumberZero :: IO ()

Function which shows message and start again

parseNumberOne :: [Char] -> IO ()

Function which checks correctness of given functions' names and if positive calls parseArgumentsName in case of bad arguments in case of 1-word commandline

parseNumberTwo :: [Char] -> IO ()

Function which checks correctness of given functions' names and if positive calls parseArgumentsName in case of bad arguments in case of 2-words commandline

parseNumberThree :: [Char] -> IO ()

Function which checks correctness of given functions' names and if positive calls parseArgumentsName in case of bad arguments in case of 3-words commandline

parseNumberFour :: [Char] -> IO ()

Function which checks correctness of given functions' names and if positive calls parseArgumentsName in case of bad arguments in case of 4-words commandline

parseNumberFive :: [Char] -> IO ()

Function which checks correctness of given functions' names and if positive calls parseArgumentsName in case of bad arguments in case of 5-words commandline